fix sap system layout rendering
