+++ web/templates/blocks/sap_system_layout.html.tmpl
@@ -9,7 +9,7 @@
                 <th scope='col'>Http port</th>
                 <th scope='col'>Https port</th>
                 <th scope='col'>Start priority</th>
-                <th scope='col'>Stauts</th>
+                <th scope='col'>Status</th>
             </tr>
         </thead>
         <tbody>
