+++ web/templates/sapsystem.html.tmpl
@@ -1,23 +1,18 @@
 {{ define "content" }}
 <div class="col">
-    {{ $EnvName := .EnvName }}
-    {{ $LandName := .LandName }}
-    {{ $Env := index .Environments .EnvName }}
-    {{ $Land := index $Env.Landscapes .LandName }}
-    {{ $SAPSystem := index $Land.SAPSystems .SAPSysName }}
-    <h6><a href="/environments">Environments</a> > <a href="/environments/{{ .EnvName }}">{{ .EnvName }}</a> > <a href="/landscapes/{{ .LandName }}?environment={{ .EnvName }}">{{ .LandName }}</a> > {{ .SAPSysName}}</h6>
+    <h6><a href="/environments">Environments</a> > <a href="/environments/{{ .Environment.Name }}">{{ .Environment.Name }}</a> > <a href="/landscapes/{{ .Landscape.Name }}?environment={{ .Environment.Name }}">{{ .Landscape.Name }}</a> > {{ .SAPSystem.Name}}</h6>
     <h1>SAP System details</h1>
     <dl class="inline">
       <dt class="inline">Name</dt>
-      <dd class="inline">{{ $SAPSystem.Name }}</dd>
+      <dd class="inline">{{ .SAPSystem.Name }}</dd>
       <dt class="inline">Type</dt>
-      <dd class="inline">{{ $SAPSystem.Type }}</dd>
+      <dd class="inline">{{ .SAPSystem.Type }}</dd>
     </dl>
     <hr/>
     <h1>Layout</h1>
-    {{ template "sap_system_layout" $SAPSystem }}
+    {{ template "sap_system_layout" .SAPSystem }}
     <hr/>
     <h1>Nodes</h1>
-    {{ template "hosts_table" $SAPSystem }}
+    {{ template "hosts_table" .SAPSystem }}
 </div>
 {{ end }}
